chore: rename row-wrapper filename
